+++ resources/database/migrations/2023_12_29_110250_create_products_table.php
@@ -5,7 +5,7 @@ use Framework\Database\Migration;
 
 return new class extends Migration
 {
-    public function up(): void
+    public function run(): void
     {
         Database::query(
             'CREATE TABLE products (' .